Add files via upload
Implementation of a simple neural network to classify grayscale images of handwritten digits (0-9) from the MNIST dataset. The dataset contains 60,000 training images and 10,000 testing images of handwritten digits, 0 - 9.

For starters, neural network with a single hidden layer and cross entropy loss was implemented and trained with provided data set. Sigmoid function is used for activation for the hidden layer and softmax function was uset for the ouput layer.
Code provides forward-propagation and back-propagation for the loss function. (regularization term was added)